[SW-2089] Expose only H2OFrame, hide remaining internal API
