+++ core/src/main/scala/ai/h2o/sparkling/H2OColumnType.scala
@@ -15,12 +15,12 @@
 * limitations under the License.
 */
 
-package ai.h2o.sparkling.job
+package ai.h2o.sparkling
 
-object H2OJobStatus extends Enumeration {
-  val DONE, CANCELLED, FAILED, RUNNING = Value
+object H2OColumnType extends Enumeration {
+  val enum, string, int, real, time, uuid = Value
 
-  def fromString(status: String): Value = {
-    values.find(_.toString == status).getOrElse(throw new RuntimeException(s"Unknown H2O's Job status $status"))
+  def fromString(dataType: String): Value = {
+    values.find(_.toString == dataType).getOrElse(throw new RuntimeException(s"Unknown H2O's Data type $dataType"))
   }
 }
