+++ core/src/main/scala/ai/h2o/sparkling/H2OFrame.scala
@@ -15,16 +15,15 @@
 * limitations under the License.
 */
 
-package ai.h2o.sparkling.frame
+package ai.h2o.sparkling
 
 import java.text.MessageFormat
 
-import ai.h2o.sparkling.backend.NodeDesc
 import ai.h2o.sparkling.backend.utils.RestApiUtils._
 import ai.h2o.sparkling.backend.utils.{RestCommunication, RestEncodingUtils}
+import ai.h2o.sparkling.backend.{H2OChunk, H2OJob, NodeDesc}
 import ai.h2o.sparkling.extensions.rest.api.Paths
 import ai.h2o.sparkling.extensions.rest.api.schema.{FinalizeFrameV3, InitializeFrameV3}
-import ai.h2o.sparkling.job.H2OJob
 import ai.h2o.sparkling.utils.Base64Encoding
 import org.apache.spark.h2o.{H2OConf, H2OContext}
 import water.api.schemas3.FrameChunksV3.FrameChunkV3
@@ -34,7 +33,7 @@ import water.api.schemas3._
 /**
  * H2OFrame representation via Rest API
  */
-class H2OFrame private(val frameId: String, val columns: Array[H2OColumn], val chunks: Array[H2OChunk])
+class H2OFrame private(val frameId: String, val columns: Array[H2OColumn], private[sparkling] val chunks: Array[H2OChunk])
   extends Serializable with RestEncodingUtils {
   private val conf = H2OContext.ensure("H2OContext needs to be running in order to create H2OFrame").getConf
   val columnNames: Array[String] = columns.map(_.name)
