+++ core/src/main/scala/ai/h2o/sparkling/backend/H2ORDD.scala
@@ -19,9 +19,8 @@ package ai.h2o.sparkling.backend
 
 import java.lang.reflect.Constructor
 
-import ai.h2o.sparkling.SparkTimeZone
 import ai.h2o.sparkling.backend.utils.{ConversionUtils, ProductType}
-import ai.h2o.sparkling.frame.H2OFrame
+import ai.h2o.sparkling.{H2OFrame, SparkTimeZone}
 import org.apache.spark.h2o.H2OContext
 import org.apache.spark.{Partition, TaskContext}
