+++ core/src/main/scala/ai/h2o/sparkling/backend/H2OSparkEntity.scala
@@ -17,7 +17,7 @@
 
 package ai.h2o.sparkling.backend
 
-import ai.h2o.sparkling.frame.H2OFrame
+import ai.h2o.sparkling.H2OFrame
 import org.apache.spark.Partition
 
 /**
