+++ core/src/main/scala/ai/h2o/sparkling/backend/Reader.scala
@@ -23,7 +23,6 @@ import ai.h2o.sparkling.backend.converters.TimeZoneConversions
 import ai.h2o.sparkling.backend.utils.SupportedTypes
 import ai.h2o.sparkling.backend.utils.SupportedTypes._
 import ai.h2o.sparkling.extensions.serde.ChunkAutoBufferReader
-import ai.h2o.sparkling.frame.H2OChunk
 import org.apache.spark.h2o.H2OConf
 import org.apache.spark.unsafe.types.UTF8String
