+++ ml/src/main/scala/ai/h2o/sparkling/ml/algos/H2OAlgorithm.scala
@@ -16,12 +16,12 @@
 */
 package ai.h2o.sparkling.ml.algos
 
+import ai.h2o.sparkling.H2OFrame
 import ai.h2o.sparkling.backend.exceptions.RestApiCommunicationException
 import ai.h2o.sparkling.backend.utils.RestCommunication
-import ai.h2o.sparkling.frame.H2OFrame
+import ai.h2o.sparkling.ml.backend.H2OModel
 import ai.h2o.sparkling.ml.models.{H2OMOJOModel, H2OMOJOSettings}
 import ai.h2o.sparkling.ml.params.H2OAlgoCommonParams
-import ai.h2o.sparkling.model.H2OModel
 import hex.Model
 import org.apache.spark.annotation.DeveloperApi
 import org.apache.spark.ml.Estimator
