+++ ml/src/main/scala/ai/h2o/sparkling/ml/algos/H2OAutoML.scala
@@ -16,12 +16,12 @@
 */
 package ai.h2o.sparkling.ml.algos
 
+import ai.h2o.sparkling.H2OFrame
 import ai.h2o.sparkling.backend.utils.{RestApiUtils, RestCommunication}
-import ai.h2o.sparkling.frame.H2OFrame
+import ai.h2o.sparkling.ml.backend.H2OModel
 import ai.h2o.sparkling.ml.models.{H2OMOJOModel, H2OMOJOSettings}
 import ai.h2o.sparkling.ml.params._
 import ai.h2o.sparkling.ml.utils.H2OParamsReadable
-import ai.h2o.sparkling.model.H2OModel
 import ai.h2o.sparkling.utils.ScalaUtils.withResource
 import ai.h2o.sparkling.utils.SparkSessionUtils
 import com.google.gson.{Gson, JsonElement}
