+++ ml/src/main/scala/ai/h2o/sparkling/ml/algos/H2OKMeans.scala
@@ -16,7 +16,7 @@
 */
 package ai.h2o.sparkling.ml.algos
 
-import ai.h2o.sparkling.frame.{H2OColumnType, H2OFrame}
+import ai.h2o.sparkling.{H2OColumnType, H2OFrame}
 import ai.h2o.sparkling.ml.params.H2OKMeansParams
 import hex.kmeans.KMeansModel.KMeansParameters
 import org.apache.spark.ml.util.{DefaultParamsReadable, Identifiable}
