+++ ml/src/main/scala/ai/h2o/sparkling/ml/features/H2OTargetEncoder.scala
@@ -18,10 +18,10 @@
 package ai.h2o.sparkling.ml.features
 
 import ai.h2o.sparkling.backend.utils.RestCommunication
+import ai.h2o.sparkling.ml.backend.H2OModel
 import ai.h2o.sparkling.ml.models.{H2OTargetEncoderBase, H2OTargetEncoderModel}
 import ai.h2o.sparkling.ml.params.H2OAlgoParamsHelper
 import ai.h2o.sparkling.ml.utils.EstimatorCommonUtils
-import ai.h2o.sparkling.model.H2OModel
 import ai.h2o.targetencoding._
 import org.apache.spark.h2o.H2OContext
 import org.apache.spark.ml.Estimator
