+++ ml/src/main/scala/ai/h2o/sparkling/ml/utils/EstimatorCommonUtils.scala
@@ -16,8 +16,8 @@
 */
 package ai.h2o.sparkling.ml.utils
 
+import ai.h2o.sparkling.backend.H2OJob
 import ai.h2o.sparkling.backend.utils.{RestApiUtils, RestCommunication}
-import ai.h2o.sparkling.job.H2OJob
 import hex.schemas.ModelBuilderSchema
 import org.apache.spark.h2o.H2OContext
